GUI: Fix maximum label/image size
Otherwise, it would crop/clip the outer parts of the image;
or it would steadily grow the GUI by a ~pixel on each image display.
Or it would not set the correct image size until manually resizing
the window.

This *should* be fixed, now; but maybe we need to revisit this later,
perhaps on some other platform or ...