GUI, query-index: Implement add/del tag in GUI, too
This currently seems a bit unreliable, though...
Is a tag added or not? The image shown should get updated,
but apparently this doesn't always happen. So the image border's
color, which indicates taggedness, might change only when reloading
the search results and displaying the image again, later. Or it
might not...

Perhaps more logging would shed light on what exactly is going on?

GUI:

* Add a toolbar to the Images tab, with add/del tag actions.
  They *should* be accessible via Alt-Plus/Alt-Minus, but
  as that's currently a bit unreliable, as previously explained,
  I'm currently not sure whether this works from outside of
  the Images tab, too, or at all.

* Implement several functionality to support updating
  the search results model, or rather just one row of it.

query-index:

* Make the results index j accessible via Search.Result.