query-index: Reuse CLIP model via models_store
With this, query-index.Search can now be re-instantiated in sub-seconds,
i.e., fast. (First attempt will still be slow of course. Also, it
currently needs manual closing of the DBs before it can be instantiated
again.)

It retains the property to always load query-index CLIP to cpu
(instead of to cuda if available), as before; though it has already
been discussed with upstream (ps-auxw) that this wasn't strictly
necessary.