WIP: Convert database.py, too, to enable multiple db instances
Again, use object-oriented style, with a module-level __getattr__()
fallback to forward to the new location of the code.

Work in progress: Currently crashes on as simple as:

    import database
    import config
    database.open_db()
    config.open_db()

..with a "TypeError: 'NoneType' object is not subscriptable"
in database.DB.decode_face(), as raw_face is None. (But why?
This doesn't happen with previous versions of the code
with the same DB!)

For viewing the diff, use -w (ignore white-space).