tests: Test `ft 0.6` as well & work-around precision problem
Previously, storing 0.6 in the config DB gave broken
0.6000000238418579 on load. Work-around using round().

The real problem is a loss of precision during struct.pack('<f') (float),
which gets expanded to the problematic number during unpack('<f').
That is, the value is stored as single-precision 4-byte float in the DB,
while the platform is using double-precision.

The problem goes away by using pack()/unpack() with '<d' (double),
but it seems we shouldn't change the DB lightly. (:-S?)
Besides, we don't need all that precision anyway, so round
to 4 digits on load from DB.

This of course loses even more precision, so perhaps the `ft` command
should be adjusted to round to 4 digits, too, or even just display
only 4 digits. Or perhaps it shouldn't, or needn't. ... Oh well.