tests: Fix test_text_search() code correctness via prepare_results()
It turns out that there weren't no results, but a driver loop
for Search.prepare_result() (singular) was missing. This is now
fixed by using the recently-added Search.prepare_results() (plural).

The test still fails, but with not enough white-listed result images!
(e.g., image of rabbit/owl are recognized as being in the forest.)
Correcting that is left for another commit.